Add API to send commands to a specific node or all nodes
Introduce new APIs for sending blocked, transaction and async commands to a
given node. This avoids the parsing step done in the legacy API and enables
keyless commands to be sent aswell, like for transactions.
Redirects nor retries are handled when using this API.

  void *redisClusterCommandToNode(redisClusterContext *cc, cluster_node *node..
  int redisClusterAppendCommandToNode(redisClusterContext *cc, cluster_node *node..
  int redisClusterAsyncCommandToNode(redisClusterAsyncContext *acc, cluster_node *node..

To send commands to all nodes a node iterator is introduced. It is based on the
dict iterator to traverse the existing nodes dict, but with the additional
handling of changes in the nodes dict.
When a route refresh is performed nodes might be added or removed, and since the
dict is currently replaced cluster_node pointers will be invalid.

A route refresh will trigger a single restart of the now invalid iterator to
make sure all nodes receives the command.

  void initNodeIterator(nodeIterator *iter, redisClusterContext *cc);
  cluster_node *nodeNext(nodeIterator *iter);

When using CLUSTER NODES for routing updates (default) its not certain that
every master node serves slots, but when using CLUSTER SLOTS it is
(enabled by `redisClusterSetOptionRouteUseSlots()`).

Usage examples:
- see ct_specific_nodes.c  (sync, pipeline and async variants)