remove dead code
a few lines above the condition `count_part < 8` is evaluated, and the
code flow jumps to the error label if the condition evaluates to true.

the only occurence when the count_part variable is changed between this
conditional and the deleted conditional is inside the above `if`, but
there is a `continue` statement at the end of the `if` block.

Therefore the deleted code cannot be reached.