Include <sys/socket.h> for builds on the BSD platforms
Networking functions, types and macros like setsockopt,
send, recv, struct linger, SOL_SOCKET and so on do not
build on the BSD platforms if the socket header is not
included (as per man pages like send(2), setsockopt(2),
etc).  Fixed by adding it to hiutil.c which appears to
be the only location we need it.