Reduce memory used by redisClusterContext without a cluster
The cluster_node table is unused when redisClusterContext is
used with a redis-server running in non-cluster mode.  This
table has one pointer entry for every possible slot which is
wasteful in the case where we know we wont use them.  Since
the initialization code explicitly writes zeroes into every
word in this table via memset (for no reason because calloc
was used) this dirties these pages too when they're unused.

Optimize for this situation by dynamically allocating this
table like we do every other larger component in a context.
Also optimize the initialization code to not (re-)set every
field to zero right after calling calloc - this API call is
required to returned already-zeroed memory by definition.