Reduce memory used by cluster_node structure
The cluster_node structure can be arranged slightly so that
a size increase from compiler padding is avoided.  This is
achieved by moving the 32bit failure_count field to follow
the port and role integers, and reducing the size of port.
The maximum TCP/IP port number is defined to fit within a
16-bit unsigned integer (this is part of the protocol so it
will never increase).

This results in a small improvement in the clustered mode
memory footprint because the context 'table' field becomes
slightly smaller.