Use hi_calloc for memory allocations where it makes sense
Make use of hi_calloc in additional cases where the dynamic
memory being allocated is mostly/all initialized to zeroes
anyway or where integer multiplication is used to calculate
the size of the allocation (can be subject to overflow). In
all cases these are not on 'hot' code paths anyway.  For a
few cases we can do away with redundant null checks and int
return codes from redundant helper routines.