ci: resolve final lgtm.com recommendations
Remove conditional code with empty body that was being
used to avoid an unused-parameter compiler warning for
the case where HI_HAVE_BACKTRACE is not set.

The (void) pattern used here is a common alternative -
it avoids the warning, makes for slightly cleaner code
and removes the empty blocks that lgtm is recommending
against.