Improve memory efficiency (#64)
* Reduce memory used by redisClusterContext without a cluster

The cluster_node table is unused when redisClusterContext is
used with a redis-server running in non-cluster mode.  This
table has one pointer entry for every possible slot which is
wasteful in the case where we know we wont use them.  Since
the initialization code explicitly writes zeroes into every
word in this table via memset (for no reason because calloc
was used) this dirties these pages too when they're unused.

Optimize for this situation by dynamically allocating this
table like we do every other larger component in a context.
Also optimize the initialization code to not (re-)set every
field to zero right after calling calloc - this API call is
required to returned already-zeroed memory by definition.

* Reduce memory used by cluster_node structure

The cluster_node structure can be arranged slightly so that
a size increase from compiler padding is avoided.  This is
achieved by moving the 32bit failure_count field to follow
the port and role integers, and reducing the size of port.
The maximum TCP/IP port number is defined to fit within a
16-bit unsigned integer (this is part of the protocol so it
will never increase).

This results in a small improvement in the clustered mode
memory footprint because the context 'table' field becomes
slightly smaller.

* Remove the pattern of doing hi_malloc followed by memset

Use hi_calloc for these cases.  It provides libc the best
opportunity to optimize this allocation, and is generally
a safer pattern from a security POV as it avoids integer
multiplication (which can be subject to overflow).

* Use hi_calloc for memory allocations where it makes sense

Make use of hi_calloc in additional cases where the dynamic
memory being allocated is mostly/all initialized to zeroes
anyway or where integer multiplication is used to calculate
the size of the allocation (can be subject to overflow). In
all cases these are not on 'hot' code paths anyway.  For a
few cases we can do away with redundant null checks and int
return codes from redundant helper routines.

* Resolve clang and test issues reported in CI with memory changes