Move SSL support to own library (#80)
* Move SSL to a separate src and headerfile

This enables us to put the SSL code in a separate library,
like its done in hiredis. A user will be able to link to the new
library when SSL is wanted, and there is no longer the need of
having `SSL_SUPPORT` defined.

* Use later version than v1.0.2 of the dependency hiredis

The tagged versions of hiredis does not include a required
packaging correction. This correction is needed to be able to
import hiredis CMake package.

* Create separate ssl library via CMake

A new target and package is created in CMake to build
the library `hiredis_cluster_ssl`

* Create separate ssl library via Make

* Update CMake configs for tests

Use the new `hiredis_library_ssl` when linking the test binaries,
but only when building with SSL/TLS enabled.

Since CMake supports transitive dependencies the `hiredis_cluster` library
will make sure `hiredis` is include when linking, thus removing the
explicit linkage in CMake.

* Update examples to use the redis_cluster_ssl library

Includes some cleanups and the use of CMAKE_PREFIX_PATH
to remove some CMake configs.

* Update README regarding the new SSL library

* Support current hiredis releases in CMake

Add the missing hiredis targets if needed. This was corrected
in hiredis via: https://github.com/redis/hiredis/pull/1005