Fix crash in async node cleanup
When a node is removed all registered command callbacks are
called with a NULL response. This can happen when a new cluster
topology is fetched. During this command callback handling a NULL
response might trigger the use of the nodes dict that is currently
under deletion, which triggers a crash.
By making sure the nodes dict is replaced before we delete the old
one a crash is avoided.