Handle removal of nodes when in a callback
Make sure we disconnect the deleted cluster node from the hiredis
async context so it wont attempt to modify the node object after
it has been deleted. This happens when a hiredis context is freed
in a callback, where the deletion is deferred to after the callback
is finished. This can happen during handling of a MOVED response.