Fix async api issues (#107)
* Fix crash in async node cleanup
When a node is removed all registered command callbacks are
called with a NULL response. This can happen when a new cluster
topology is fetched. During this command callback handling a NULL
response might trigger the use of the nodes dict that is currently
under deletion, which triggers a crash.
By making sure the nodes dict is replaced before we delete the old
one a crash is avoided.

* Handle redisClusterAsyncCommand() during disconnect
Return an error instead of asserting when a user attempts to
send a command while a connection is disconnecting.

* Handle removal of nodes when in a callback
Make sure we disconnect the deleted cluster node from the hiredis
async context so it wont attempt to modify the node object after
it has been deleted. This happens when a hiredis context is freed
in a callback, where the deletion is deferred to after the callback
is finished. This can happen during handling of a MOVED response.

Co-authored-by: Viktor Söderqvist <viktor.soderqvist@est.tech>