Allow callbacks in context when node already removed
When a context is freed all callbacks are called with a NULL response.
During the callback handling there is an attempt to update the
topology after a number of attempts. If the node is removed during
this topology change the context's pointer to the node will be cleared.
The following callbacks needs to finish and not asserting.