Find hiredis libraries when CMake packages not found
The CMake build attempts to find the installed CMake packages for hiredis,
which is only installed if hiredis was built and installed using CMake.
By adding an alternative search for the library files we now handle
setups where hiredis was installed using its regular makefile.

Added a build example for verification in CI.