Use correct key format during use of CLUSTER NODES
node->addr is used as key in the nodes dict in the cluster context.
During ASK redirects the key is also expected to be ip:port and this
change can fix a problem of creating node doublets is some scenarios.

Both the add node api and CLUSTER SLOTS also uses ip:port as the key.

The key could previously be:
 127.0.0.1:30004@31004,,shard-id=69bc080733d1355567173199cff4a6a039a2f024
instead of:
 127.0.0.1:30004