Use the async api when a connection error triggers a slot update (#144)
When a command response indicates a communication error the slot map
is updated. This is now updated using the async api, which avoids
blocking calls to connect when querying the cluster node.

Primarily selects a connected node found close to a randomly picked index of
all known nodes. The random index should give a more even distribution of
selected nodes. If no connected node is found while iterating to this index
the remaining nodes are also checked until a connected node is found.
If no connected node is found a node for which a connect has not been attempted
within throttle-time, and is found near the picked index, is selected.

* Remove use of Redis config cluster-node-timeout
  The attempt to update the slot map will be immediate.

* Update slot map immediately for communication problems
  Make sure we throttle the slot map update to 1/sec after communications errors.
  Update testcase timeouts to include slotmap updates.

* Fix redisClusterAsyncDisconnect()
  Let hiredis cleanup after a disconnect like it does after a failure,
  i.e call unlinkAsyncContextAndNode().
  This fixes a thread-sanitizer issue.

* Add lastConnectionAttempt to redisClusterNode
  Add a timestamp to each redisClusterNode to indicate when
  the last connection attempt was performed to this node.