Fix the check for ":0@0" in CLUSTER NODES result (#199)
Fix the check for the `ip:port@cport` part when parsing a line in the CLUSTER NODES result. The current implementation assumes that the entire part is only 2 characters when the IP and port are not set (IP is empty and port is 0). However, this is incorrect because the `@cport` is also present in Redis 4+ so the current check fails. Additionally, there may be an optional `,hostname` as well which is not accounted for.