Fix a compilation warning in command.c relating to kpos shadowing
When building with the gcc -Wshadow option, a warning is generated
in redis_parse_cmd() because the second and third declarations of
the same-named kpos variable share scope.  It's benign but it'd be
great to have clean compilation when using this option.

This commit implements Viktor Söderqvist's refactoring approach as
described in github PR #201.